tests: extract test helpers into commons
Currently, test helpers functions and actual test logics are put into the same file. The file will get much longer as new integration tests are added in the future. Solution is to extract the common helpers to a common submodule so future integration tests can reuse them.

This commit does not add or remove any tests, only the organization.
