+++ lightway-core/tests/common/data/ca_cert_der_2048
